requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
aniso8601==9.0.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.11.1
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
distlib==0.3.8
dnspython==2.6.1
dominate==2.9.1
email_validator==2.2.0
et-xmlfile==1.1.0
exrex==0.11.0
filelock==3.13.1
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
MouseInfo==0.1.3
multidict==6.0.5
numpy==1.26.4
openpyxl==3.0.10
packaging==24.0
pandas==2.2.1
picozero==0.4.2
pillow==10.2.0
platformdirs==4.1.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2024.1
pyusb==1.2.1
requests==2.28.1
requests-futures==1.0.1
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.30
stem==1.8.2
thorndyke==2022.1.2.1
torrequest==0.1.0
tqdm==4.66.2
typing_extensions==4.12.1
tzdata==2024.1
urllib3==1.26.18
virtualenv==20.25.0
visitor==0.1.3
Werkzeug==3.0.3
WTForms==3.1.2
yarl==1.9.4
